Moved PermissionConsts to core/permissions/Permission
