+++ src/components/permissions/PermissionsPanel.js
@@ -3,7 +3,7 @@ import Dropdown from 'react-dropdown';
 import Select from 'react-select';
 import { PermissionsApi, UsersApi } from 'loom-data';
 import StringConsts from '../../utils/Consts/StringConsts';
-import PermissionsConsts from '../../utils/Consts/PermissionsConsts';
+import PermissionsConsts from '../../core/permissions/Permission';
 import UserRoleConsts from '../../utils/Consts/UserRoleConsts';
 import styles from './permissions.module.css';
 import Utils from '../../utils/Utils';
